Update SecurityCheck
I have updated PassGuard.SecurityCheck to return error codes alongside the result, to allow the frontend to handle errors gracefully instead of defaulting to a secure credential state. Alongside this, I have allowed the algorithm to fetch the HIBP description, and have used regexp to sanitise the input of HTML tags.

Improved comments throughout this algorithm.