Create PasswordGenerator
An algorithm that can create passwords, with a random character set, a word dictionary, or a letter only system. If the Type is 1, the word generation system is used. If the Type is 2 the character generation system is used, and if the Type is 3 the character generation system is used, but special characters are excluded.

I modularised parts of this, so that the logic for choosing the characters/words to use can be easily swapped out.